refactor: changed navbar logic to reduce duplication
