+++ src/app/(internal-sites)/todo/page.tsx
@@ -1,0 +1,11 @@
+import Navbar from "~/components/navbar";
+
+const Todo = () => {
+  return (
+    <div className="flex h-full flex-col items-center justify-center lg:pl-24">
+      <p>Todo</p>
+    </div>
+  );
+};
+
+export default Todo;
